A bit of refactoring to clean up the code and abstractions a bit
